Refactor tracking so that utility functionality can be used independently from motile
